ARM: dts: Add am335x mcasp with l3 data port ranges
* Peter Ujfalusi <peter.ujfalusi@ti.com> [181210 17:00]:
> On 10/12/2018 18.38, Tony Lindgren wrote:
> > OK so to prevent the hwmod code from resetting gpios very
> > early, this now might become necessary. Care to try with
> > this also applied?
>
> With this applied there is life on the console, however it ends up badly:

OK thanks for testing. I'll take a look on how to skip
early reset for omap_hwmod.c for the dts case instead of
dropping the sysc platform data.

For the LCDC, does the following revert fix it? If so,
I need to check the flags for am33xx_lcdc_hwmod
HWMOD_SWSUP_SIDLE | HWMOD_SWSUP_MSTANDBY.

Regards,

Tony

8< -------------------------