arm64: dts: ti: j721e-common-proc-board: Analog audio support
The codec is wired in multi DAI setup (DIN1/2/3/4/DOUT1/2/3 is connected to
McASP serializer).

For some reason with DIN1/DOUT1 setup we can not get 96KHz audio to work as
we would need to run REFCLK2 in 256xFS rate, but then in all sampling rates
the channel 4-7 is going to be muted :o

For other unknown reason 256xFS SCKI does not have this quirk and we can
use 96KHz as well.

There are couple of notable difference compared to DIN1/DOUT1 mode:
the channel mapping is 'random' for first look compared to the single
serializer setup.
The McASP driver does not have support for right_j currently which would
allow us to use 16 bit samples as well (it is supported by pcm3158a only if
it is slave and non TDM mode is used).

So the channel mapping:
        _      _      _
       |o|c1  |o|p1  |o|p3
 _     | |    | |    | |
|o|c3  |o|c2  |o|p4  |o|p2
------------------------

c1/2/3 - capture jacks (3rd is line)
p1/2/3/4 - playback jacks (4th is line)

2 channel audio (stereo):
0 (left):  p1/c1 left
1 (right): p1/c1 right

4 channel audio:
0: p1/c1 left
1: p2/c2 left
2: p1/c1 right
3: p2/c2 right

6 channel audio
0: p1/c1 left
1: p2/c2 left
2: p3/c3 left
3: p1/c1 right
4: p2/c2 right
5: p3/c3 right

8 channel audio
0: p1/c1 left
1: p2/c2 left
2: p3/c3 left
3: p4 left
4: p1/c1 right
5: p2/c2 right
6: p3/c3 right
7: p4 right

The reason is that we need to put data to all enabled serializers first.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>