BAD!! drm/bridge: cdns-mhdp: Filter modes according to available bandwidth BAD!!!
Do link training in mode_fixup() call-back to get the available
bandwidth. Return false if the mode exceeds tho link bandwidth and
power down the link, since there is no disable call after a failed
fixup. The mode_fixup is called in the check phase, so this change is
needed for us to be able to reject bad configuration, before commit -
which should not fail.

Signed-off-by: Jyri Sarha <jsarha@ti.com>