ARM: dts: k3-j721e-common-proc-board: enable DP0 power
DP connector needs to provide 3.3V on the connector's power pin. Devices
like active adapters need the power to function.

This patch enables the power with a GPIO hog, so it's always enabled as
the power needs to be enabled for us to even detect that an active
adapter is connected.

Probably the power should be enabled only if the DP driver is loaded,
and disabled on remove or suspend, but as it's not clear where the power
should be handled (it's not really part of the DP IP in any way), we use
GPIO hog for now.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Acked-by: Andrew F. Davis <afd@ti.com>