arm64: dts: ti: k3-j721e-common-proc-board: Switch to custom audio support
Due to the audio setup in hardware with the generic sound card the features
we can support is limited.

With the custom audio driver it is possible to support both 48KHz and
44.1Khz sampling rate families by switching between PLL4/15 as needed.

We can also support both 16 and 24 bit audio in this case as with a custom
driver we can change the slot_width on the bus.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>