drm/bridge: cdns-mhdp: Check link status if HPD interrupt is received
Add code to interrupt handler to check if the DisplayPort link is
still up if it is supposed to be. First check that the HPD state is
asserted (the connector is plugged). Then read DPCD status and check
if the channel eq and clock recovery is still ok. Mark the connector's
link status property as bad if DPCD read fails or link status is not
ok. The link status check and the property setting is done while
holding mode_config.mutex to avoid a race. The fine grained locking
implementation in drm_connector_set_link_status_property() is broken
so we must use the legacy giant-lock.

Signed-off-by: Jyri Sarha <jsarha@ti.com>