drm/bridge: cdns-mhdp: Check link status if HPD interrupt is received
Add code to interrupt handler to check if the DisplayPort link is
still up if is supposed to be. First check that the HPD state is
asserted (the connector is plugged). Then read DPCD status and check
if the channel eq and clock recovery is still ok. Mark the connectors
link status property as bad if DPCD read fails or link status is not
ok.

Signed-off-by: Jyri Sarha <jsarha@ti.com>