drm/bridge: Remove all traces of cdns-mhdp of tc358768 driver
This commit resets the drm bridge driver to a state where it does not
have any drivers added in ti-linux-4.19.y, when compared to v4.19.73
stable version. However, all the changes to dumb-vga-dac, sii902x,
tc358767, and ti-tfp410, are kept since part of them is needed for
patches we have back ported from mainline DRM core. But namely all
changes associated to cdns-mhdp and tc358768 drivers are lost.

Signed-off-by: Jyri Sarha <jsarha@ti.com>