phy: add phy_get_link_rate()/phy_set_link_rate() calls
This adds a pair of APIs that allows the generic PHY subsystem to
provide information on the PHY link rate. The PHY provider driver may
use phy_set_link_rate() to set the link rate that the PHY supports.
The controller driver may then use phy_get_link_rate() to fetch the
PHY link rate in order to properly configure the controller.

Signed-off-by: Yuti Amonkar <yamonkar@cadence.com>